requirements.txt
# Web Framework & Reactivity
Flask==3.1.1
Flask-WTF==1.2.2

# Data Layer
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.4
Flask-Migrate==4.1.0
supabase==2.13.0

# Auth & Security
Werkzeug==3.1.3
python-dotenv==1.0.1
email-validator==2.2.0
Flask-Login==0.6.3

# Production
gunicorn==23.0.0
gevent==24.11.1
